feat: dump 3-dimensional data to ply to visualize_
Now if the output of the model have more than 3 di
mensions, we can use PCA to extract the first 3 di
mensions as xyz coordinates to visualize the data.
But original DrLim seems not perform well on Cifar
10 dataset, may be due to the rgb channels confuse
Add the train and test scripts, so train and  test
can be launch instead of debug to  accelerate  the
process.