feat: add triplet loss comparison_________________
However, the performance of triplet loss on Cifar1
0 dataset is not good, may due to  the  simplicity
of base network model. I may consider seperate the
loss and network model, and use  a  more  complex,
deeper network. Since images in Cifar10  not  only
have target, but also contain a background,  which
may cause some noise feature. I have also tried to
feed the model with gray-scale Cifar10 images, but
not much performance gained...
Main changes of this commit also include a optimiz
on loss to centeralized the manifold coordinates.